ROC-8038 | added new event and updated code for bulk print transfer
