--- src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/transfercase/CoverLetterGenerator.java
@@ -1,59 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker.transfercase;
-
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
-import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.DocAssemblyService;
-import uk.gov.hmcts.cmc.claimstore.services.ccd.legaladvisor.DocAssemblyTemplateBody;
-import static uk.gov.hmcts.cmc.claimstore.utils.DocumentNameUtils.buildNoticeOfTransferForCourtFileBaseName;
-import static uk.gov.hmcts.cmc.claimstore.utils.DocumentNameUtils.buildNoticeOfTransferForDefendantFileBaseName;
-
-@Component
-public class CoverLetterGenerator {
-
-    private final DocAssemblyService docAssemblyService;
-    private final NoticeOfTransferLetterTemplateMapper noticeOfTransferLetterTemplateMapper;
-
-    public CoverLetterGenerator(
-        DocAssemblyService docAssemblyService,
-        NoticeOfTransferLetterTemplateMapper noticeOfTransferLetterTemplateMapper) {
-        this.docAssemblyService = docAssemblyService;
-        this.noticeOfTransferLetterTemplateMapper = noticeOfTransferLetterTemplateMapper;
-    }
-
-    public CCDDocument generate(CCDCase ccdCase,
-                                String authorisation,
-                                NoticeOfTransferLetterType noticeOfTransferLetterType,
-                                String templateId) {
-        DocAssemblyTemplateBody formPayloadForCourt =
-            noticeOfTransferLetterTemplateMapper.noticeOfTransferLetterBodyForCourt(ccdCase, authorisation);
-
-        return docAssemblyService.generateDocument(ccdCase,
-            authorisation,
-            formPayloadForCourt,
-            templateId)
-            .toBuilder()
-            .documentFileName(buildNoticeOfTransferLetterFileName(ccdCase, noticeOfTransferLetterType))
-            .build();
-    }
-
-    private String buildNoticeOfTransferLetterFileName(
-        CCDCase ccdCase,
-        NoticeOfTransferLetterType noticeOfTransferLetterType
-    ) {
-        String basename;
-
-        switch (noticeOfTransferLetterType) {
-            case FOR_COURT:
-                basename = buildNoticeOfTransferForCourtFileBaseName(ccdCase.getPreviousServiceCaseReference());
-                break;
-            case FOR_DEFENDANT:
-                basename = buildNoticeOfTransferForDefendantFileBaseName(ccdCase.getPreviousServiceCaseReference());
-                break;
-            default:
-                throw new IllegalArgumentException(noticeOfTransferLetterType
-                    + " noticeOfTransferLetterType unable to be handled");
-        }
-        return String.format("%s.pdf", basename);
-    }
-}
