+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/transfercase/TransferCasePostProcessor.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker.transfercase;
 
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.CallbackParams;
@@ -10,22 +9,17 @@ import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 
-import java.time.LocalDate;
-
 @Service
 public class TransferCasePostProcessor {
 
     private final CaseDetailsConverter caseDetailsConverter;
-    private final TransferCaseNotificationsService transferCaseNotificationsService;
-    private final TransferCaseDocumentPublishService transferCaseDocumentPublishService;
+    private final BulkPrintTransferService bulkPrintTransferService;
 
     public TransferCasePostProcessor(
         CaseDetailsConverter caseDetailsConverter,
-        TransferCaseNotificationsService transferCaseNotificationsService,
-        TransferCaseDocumentPublishService transferCaseDocumentPublishService) {
+        BulkPrintTransferService bulkPrintTransferService) {
         this.caseDetailsConverter = caseDetailsConverter;
-        this.transferCaseNotificationsService = transferCaseNotificationsService;
-        this.transferCaseDocumentPublishService = transferCaseDocumentPublishService;
+        this.bulkPrintTransferService = bulkPrintTransferService;
     }
 
     public CallbackResponse completeCaseTransfer(CallbackParams callbackParams) {
@@ -34,36 +28,10 @@ public class TransferCasePostProcessor {
         CCDCase ccdCase = caseDetailsConverter.extractCCDCase(caseDetails);
         Claim claim = caseDetailsConverter.extractClaim(caseDetails);
         String authorisation = callbackParams.getParams().get(CallbackParams.Params.BEARER_TOKEN).toString();
-
-        ccdCase = transferCaseDocumentPublishService.publishCaseDocuments(ccdCase, authorisation, claim);
-
-        sendEmailNotifications(ccdCase, claim);
-
-        ccdCase = updateCaseData(ccdCase);
-
+        ccdCase = bulkPrintTransferService.transferCase(ccdCase, claim, authorisation);
         return AboutToStartOrSubmitCallbackResponse
             .builder()
             .data(caseDetailsConverter.convertToMap(ccdCase))
             .build();
-    }
-
-    private CCDCase updateCaseData(CCDCase ccdCase) {
-
-        return ccdCase.toBuilder()
-            .transferContent(ccdCase.getTransferContent().toBuilder().dateOfTransfer(LocalDate.now()).build())
-            .build();
-    }
-
-    private void sendEmailNotifications(CCDCase ccdCase, Claim claim) {
-
-        transferCaseNotificationsService.sendClaimUpdatedEmailToClaimant(claim);
-
-        if (isDefendantLinked(ccdCase)) {
-            transferCaseNotificationsService.sendClaimUpdatedEmailToDefendant(claim);
-        }
-    }
-
-    private boolean isDefendantLinked(CCDCase ccdCase) {
-        return !StringUtils.isBlank(ccdCase.getRespondents().get(0).getValue().getDefendantId());
     }
 }
